WIP Some progress: Login half-working, remove unnecessary context
so when I click login and when the `/callback` happens with the code,
the app doesn't use that therefore, no login is registered. But if I
maintain the url state, and re-click the login button, the user is
logged in (because we have the code in the url) and the sdk also clears
the url for me. However, on refresh, the app doesn't remember that it
has the access token (local storage), on clicking the login buton again,
the user is logged in instantly because we already had the access tokens
in the local storage so the app uses those and does NOT make the
redirect request.

HOWEVER, I think I could have done it much simply. Remove the contexts
and use the modified useSpotify hook.